backend/requirements.txt
fastapi==0.128.4
uvicorn==0.40.0
python-multipart==0.0.22
python-dotenv==1.2.1
openai-whisper==20250625
torch==2.10.0
torchaudio==2.10.0
moviepy==2.2.1
numpy>=1.25.0,<2.4
